requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
langchain
langchain-community
pypdf
tiktoken
openai
pinecone
